+++ 3Erronka/src/Modelo/.gitignore
@@ -1,0 +1,16 @@
+/Aretoa.class
+/AretoaTest.class
+/Bezeroa.class
+/BezeroaTest.class
+/Erosketa.class
+/ErosketaTest.class
+/Karteldegia.class
+/KarteldegiaTest.class
+/Pelikula.class
+/PelikulaTest.class
+/Saioa.class
+/SaioaTest.class
+/Sarrera.class
+/SarreraTest.class
+/Zinema.class
+/ZinemaTest.class
