+++ 3Erronka/src/Testak/.gitignore
@@ -1,0 +1,8 @@
+/AretoaTest.class
+/BezeroaTest.class
+/ErosketaTest.class
+/KarteldegiaTest.class
+/PelikulaTest.class
+/SaioaTest.class
+/SarreraTest.class
+/ZinemaTest.class
