+++ 3Erronka/src/Testak/AretoaTest.java
@@ -1,9 +1,13 @@
-package Modelo;
+package Testak;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 import org.junit.BeforeClass;
 import org.junit.Test;
+
+import Modelo.Aretoa;
 
 public class AretoaTest {
