+++ 3Erronka/src/Testak/BezeroaTest.java
@@ -1,9 +1,11 @@
-package Modelo;
+package Testak;
 
 import static org.junit.Assert.*;
 
 import org.junit.BeforeClass;
 import org.junit.Test;
+
+import Modelo.Bezeroa;
 
 public class BezeroaTest {
